BOOTSTRAP " portfolio "
In this HTML document, linked with BOOTSTRAP, no use of CSS (forbidden technique) and for this assignment we had to create the appearance of a website so that we can discuss it in more detail and more interactively later in another exercise. For now, I'm very happy with the results I've had so far and can't wait to continue working on this!